quick fix : Integer division in floating point context

+++ src/main/java/logisticspipes/blocks/powertile/LogisticsPowerJunctionTileEntity.java
@@ -3,7 +3,6 @@ package logisticspipes.blocks.powertile;
 import cofh.api.energy.IEnergyHandler;
 import cpw.mods.fml.common.Optional;
 import ic2.api.energy.tile.IEnergySink;
-import java.util.List;
 import logisticspipes.LPConstants;
 import logisticspipes.api.ILogisticsPowerProvider;
 import logisticspipes.blocks.LogisticsSolidTileEntity;
@@ -29,6 +28,8 @@ import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.ForgeDirection;
+
+import java.util.List;
 
 @Optional.InterfaceList({
     @Optional.Interface(modid = "IC2", iface = "ic2.api.energy.tile.IEnergySink"),
@@ -341,8 +342,7 @@ public class LogisticsPowerJunctionTileEntity extends LogisticsSolidTileEntity
         }
         transferFromIC2Buffer();
         // round up so we demand enough to completely fill visible storage
-        return (freeSpace() + LogisticsPowerJunctionTileEntity.IC2Multiplier - 1)
-                / LogisticsPowerJunctionTileEntity.IC2Multiplier;
+		return (double) (freeSpace() + LogisticsPowerJunctionTileEntity.IC2Multiplier - 1) / LogisticsPowerJunctionTileEntity.IC2Multiplier;
     }
 
     @Override
@@ -373,7 +373,7 @@ public class LogisticsPowerJunctionTileEntity extends LogisticsSolidTileEntity
         int RFspace = freeSpace() * LogisticsPowerJunctionTileEntity.RFDivisor - internalRFbuffer;
         int RFtotake = Math.min(maxReceive, RFspace);
         if (!simulate) {
-            addEnergy(RFtotake / LogisticsPowerJunctionTileEntity.RFDivisor);
+            addEnergy((float) RFtotake / LogisticsPowerJunctionTileEntity.RFDivisor);
             internalRFbuffer += RFtotake % LogisticsPowerJunctionTileEntity.RFDivisor;
             if (internalRFbuffer >= LogisticsPowerJunctionTileEntity.RFDivisor) {
                 addEnergy(1);
